% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_menger_curvature.R
\name{calculate_menger_curvature}
\alias{calculate_menger_curvature}
\title{Menger Curvature}
\usage{
calculate_menger_curvature(interpolated_dose_vector, predicted_response_vector)
}
\arguments{
\item{interpolated_dose_vector}{Numeric vector of length 3 containing
doses corresponding to x-axis values.}

\item{predicted_response_vector}{Numeric vector of length 3 containing
responses corresponding to y-axis values}
}
\description{
Given three xy-coordinates, calculates their Menger Curvature.
}
