% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pod_quantiles.R
\name{calculate_pod_quantiles}
\alias{calculate_pod_quantiles}
\title{Point of Departure Estimation}
\usage{
calculate_pod_quantiles(
  dose_response_data,
  resample_size = 1000,
  interpolation_size = 50,
  quantile_probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{dose_response_data}{Data frame with two numeric columns. The first
column should contain doses and the second column should contain the responses.}

\item{resample_size}{The number of bootstrap samples}

\item{interpolation_size}{The number of doses to interpolate between the
minimum and maximum doses for spline prediction}

\item{quantile_probs}{The quantiles to report for POD predictions}
}
\value{
The returned object prints the POD quantiles to the console. The returned object
is a gravee list with four named components. The first component \code{pods} is a
vector of length \code{resample_size} containing the estimated PODs from all bootstrap
samples, on the original dose scale. The second element \code{pod_quantiles} contains
the POD quantiles, with boundaries defined by \code{quantile_probs}. The third
component \code{all_res} is a list of length \code{resample_size} containing the
interpolated spline predictions from each bootstrap sample and curvature
measurements along the fit curve. The fourth component \code{bs_samples} is a list
of length \code{resample_size} containing the sampled values for each bootstrap
sample.
}
\description{
Estimate a chemical Point of Departure (POD) from dose-response data using
bootstrap resampling and spline meta-regression
}
\details{
Analysis is performed on log10-transformed doses. The returned PODs are
back transformed to the original dose scale.

If the input data contains a dose = 0, the 0 dose is converted to be 1/10th
the minimum non-zero dose, such that after log10-transformation, the distance
between the 0 dose and minimum non-zero dose is 1. For PODs and interpolated
doses less than the minimum non-zero dose, values are back-transformed and
rescaled to be left bound by 0. Thus, in the output, for interpolated doses
less than the minimum non-zero dose, values in the \code{log10_dose} column
will not be equal to log10 of the \code{dose} column.

POD estimation assumes that the input dose-response data are asymptotic at
low-doses. If the lower limit of the POD credible interval is within the two
smallest doses, the function will return a warning:
\verb{POD Credible Interval violates low-dose asymptote assumption.} In this case,
verify that an asymptote has been established at lower doses of the input data.
}
\examples{
x <- rep(c(0,0.05,0.1,0.5,1), each = 3)
y <- 1/exp(-x) + abs(rnorm(length(x)))
df <- data.frame(x,y)
pods <- calculate_pod_quantiles(df)
plot(pods)
}
