% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pod_quantiles.R
\name{calculate_pod_quantiles}
\alias{calculate_pod_quantiles}
\title{Point of Departure Estimation}
\usage{
calculate_pod_quantiles(
  dose_response_data,
  resample_size = 1000,
  interpolation_size = 50,
  quantile_probs = c(0.05, 0.5, 0.95),
  spline_res = F
)
}
\arguments{
\item{dose_response_data}{Data frame with rows representing assays/genes and columns representing samples.
The first column should contain assay/gene identifiers. Column names should be numeric values of the
sample's concentration.}

\item{resample_size}{The number of resamples to perform}

\item{interpolation_size}{The number of doses to perform spline predictions for response}

\item{quantile_probs}{The quantiles to report for POD predictions}

\item{spline_res}{Logical, default = False Return the output of the spline fit}
}
\description{
Returns quantiles from estimated Point of Departure distribution
}
